+++ modules/juce_audio_formats/codecs/juce_OggVorbisAudioFormat.cpp
@@ -308,8 +308,8 @@ public:
                 if (ogg_stream_flush (&os, &og) == 0)
                     break;
 
-                output->write (og.header, og.header_len);
-                output->write (og.body, og.body_len);
+                output->write (og.header, (size_t) og.header_len);
+                output->write (og.body,   (size_t) og.body_len);
             }
 
             ok = true;
@@ -391,8 +391,8 @@ public:
                     if (ogg_stream_pageout (&os, &og) == 0)
                         break;
 
-                    output->write (og.header, og.header_len);
-                    output->write (og.body, og.body_len);
+                    output->write (og.header, (size_t) og.header_len);
+                    output->write (og.body,   (size_t) og.body_len);
 
                     if (ogg_page_eos (&og))
                         break;
