+++ modules/juce_opengl/opengl/juce_OpenGLImage.cpp
@@ -100,8 +100,8 @@ private:
 
         static void verticalRowFlip (PixelARGB* const data, const int w, const int h)
         {
-            HeapBlock<PixelARGB> tempRow (w);
-            const int rowSize = sizeof (PixelARGB) * w;
+            HeapBlock<PixelARGB> tempRow ((size_t) w);
+            const size_t rowSize = sizeof (PixelARGB) * (size_t) w;
 
             for (int y = 0; y < h / 2; ++y)
             {
@@ -122,8 +122,8 @@ private:
 
         void write (const PixelARGB* const data) const noexcept
         {
-            HeapBlock<PixelARGB> invertedCopy (area.getWidth() * area.getHeight());
-            const int rowSize = sizeof (PixelARGB) * area.getWidth();
+            HeapBlock<PixelARGB> invertedCopy ((size_t) (area.getWidth() * area.getHeight()));
+            const size_t rowSize = sizeof (PixelARGB) * (size_t) area.getWidth();
 
             for (int y = 0; y < area.getHeight(); ++y)
                 memcpy (invertedCopy + area.getWidth() * y,
@@ -142,7 +142,7 @@ private:
     struct DataReleaser  : public Image::BitmapData::BitmapDataReleaser
     {
         DataReleaser (OpenGLFrameBuffer& fb, int x, int y, int w, int h)
-            : data (w * h),
+            : data ((size_t) (w * h)),
               writer (fb, x, y, w, h)
         {}
